% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_biv_analysis.R
\name{mod_biv_analysis_server}
\alias{mod_biv_analysis_server}
\title{biv_analysis Server Function}
\usage{
mod_biv_analysis_server(input, output, session, data, genre)
}
\arguments{
\item{input, output, session}{Internal parameters for {shiny}.}

\item{data}{Reactive dataframe that is filtered based on user selected movie genre.}

\item{genre}{Reactive output of genre inputs present in app_ui.}
}
\description{
A server module to pupulate content of bivariate analysis tab.
}
