% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_biv_analysis.R
\name{mod_biv_analysis_ui}
\alias{mod_biv_analysis_ui}
\title{biv_analysis UI Function}
\usage{
mod_biv_analysis_ui(id)
}
\arguments{
\item{id}{Internal parameters for {shiny}.}
}
\description{
A shiny Module.
}
