% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_cluster.R
\name{mod_cluster_server}
\alias{mod_cluster_server}
\title{cluster Server Function}
\usage{
mod_cluster_server(input, output, session, data)
}
\arguments{
\item{input, output, session}{Internal parameters for {shiny}.}

\item{data}{Reactive dataframe that is filtered based on user selected movie genre.}
}
\description{
A server module that populates the cluster tab.
}
