% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_cluster.R
\name{mod_cluster_ui}
\alias{mod_cluster_ui}
\title{cluster UI Function}
\usage{
mod_cluster_ui(id)
}
\arguments{
\item{id}{Internal parameters for {shiny}.}
}
\description{
A shiny Module to populate the contents for the tab showing clustering.
}
