% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_infoBox.R
\name{mod_infoBox_server}
\alias{mod_infoBox_server}
\title{infoBox Server Function}
\usage{
mod_infoBox_server(input, output, session, title, icon, data, based_on, target)
}
\arguments{
\item{input, output, session}{Internal parameters for {shiny}}

\item{title}{Title to be used in the info box}

\item{icon}{Icon to be used in the box}

\item{data}{Reactive data frame to be used to populate the content inside the box}

\item{based_on}{Based on which factor to calculate the best performer}

\item{target}{Which information about the top performer to collect}
}
\description{
Server function to populate information boxes created by mod_infoBox_ui().
}
