% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_infoBox.R
\name{mod_infoBox_ui}
\alias{mod_infoBox_ui}
\title{infoBox UI Function}
\usage{
mod_infoBox_ui(id)
}
\arguments{
\item{id, input, output, session}{Internal parameters for {shiny}.}
}
\description{
Ui function to create the information boxes showing top performers at the top of the Summary tab.
}
