% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_valueBox.R
\name{mod_valueBox_server}
\alias{mod_valueBox_server}
\title{valueBox Server Function}
\usage{
mod_valueBox_server(input, output, session, subtitle, icon, data, target)
}
\arguments{
\item{subtitle}{Title to be used in the info box}

\item{icon}{Icon to be used in the box}

\item{data}{Reactive data frame to be used to populate the content inside the box}

\item{target}{Which information about the top performer to collect}

\item{id, input, output, session}{Internal parameters for {shiny}}
}
\description{
Server function to populate value boxes created by mod_valueBox_ui().
}
