% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_valueBox.R
\name{mod_valueBox_ui}
\alias{mod_valueBox_ui}
\title{valueBox UI Function}
\usage{
mod_valueBox_ui(id)
}
\arguments{
\item{id}{Internal parameters for {shiny}.}
}
\description{
Ui function to create the value boxes showing performance at the bottom of the Summary tab.
}
